#83 - [Style] PressShutterViewController 화면 구성
